186 refactor settingscontentcomponent (#187)
* feat: made OrganizationSetupCOmponent

* feat: split settingscontent into multiple smaller components

* feat: extracted out styling from PAT component

* feat: Changed PAT component to only us MUI

* chore: removed unnecesarry styling

* chore: removed previous commit

* fix: replaced aksel icons with MUI

---------
